+++ Source/YaOpt/Helpers/ListerThingsHelper.cs
@@ -1,7 +1,7 @@
+using HarmonyLib;
 using RimWorld;
 using System.Collections.Generic;
 using System.Runtime.CompilerServices;
-using HarmonyLib;
 using Verse;
 using static YaOpt.Helpers.ListerThingsIndexer;